+++ src/test/java/test/assertk/assertions/JavaNullableString.java
@@ -1,0 +1,8 @@
+package test.assertk.assertions;
+
+public class JavaNullableString {
+
+    public static String string() {
+        return null;
+    }
+}